% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_structure.R
\name{print.odk_structure}
\alias{print.odk_structure}
\title{Print an odk_structure object}
\usage{
\method{print}{odk_structure}(x, ...)
}
\arguments{
\item{x}{An \code{odk_structure} object returned by \code{\link[=detect_structure]{detect_structure()}}.}

\item{...}{Ignored.}
}
\description{
Print an odk_structure object
}
